Add files via upload
The radiative equilibrium protomodel of the Sun's atmosphere by Karl Schwarzschild tabulated in Schwarzschild (1906) in Table on the page 47 and in Loeser (1966) in Table on the page 30.

References:

Schwarzschild, K.: On the equilibrium of the Sun’s atmosphere. Nachrichten von
der königlichen gesellschaft der wissenschaften zu Göttingen. Math.-phys. Klasse
195, 41-53 (1906)
https://ui.adsabs.harvard.edu/abs/1906WisGo.195...41S

Loeser, R.: On the equilibrium of the sun’s atmosphere. In: Menzel, D.H. (ed.)
Selected Papers on the Transfer of Radiation, pp. 25-34. Dover Publications, Inc.,
New York (1966)